services/market-ingestor/requirements.txt
fastapi
uvicorn
pydantic
pandas
yfinance
kafka-python
python-dateutil
requests
kafka-python
yfinance
websockets
pytz
